packages/taiko-client/tools/requirements.txt
web3==6.15.1
python-dotenv==1.0.1
argparse==1.4.0